+++ CHANGELOG.md
@@ -32,11 +32,14 @@
   * it is now a component of `MonadBlockChainBalancing`
   * it can be turned on/off in pretty-printing options
   * it now displays the discarding of utxos during balancing.
+  * it now displays when the user specifies useless collateral utxos.
   * it is not visible from outside of `cooked-validators`
 
 ### Fixed
 
 - All kinds of scripts can now be used as reference scripts.
+- Transactions that do not involve scripts are now properly generated without any
+  collateral.
 
 ## [[4.0.0]](https://github.com/tweag/cooked-validators/releases/tag/v4.0.0) - 2024-06-28